pyproject.toml
[tool.poetry]
name = "test task rlt"
version = "0.1.0"
description = ""
authors = ["Arthur-king18 <arturik060406@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "3.11"
dnspython = "^2.4.2"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
aiogram = "2.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
